Move main into it's own file.
